Add support for filling backed enum types in the DB ORM layer, and cleanup some formatting issues
