+++ lib/EbookSource.php
@@ -1,16 +1,4 @@
 <?
-enum EbookSourceType{
-	case ProjectGutenberg;
-	case ProjectGutenbergAustralia;
-	case ProjectGutenbergCanada;
-	case InternetArchive;
-	case HathiTrust;
-	case Wikisource;
-	case GoogleBooks;
-	case FadedPage;
-	case Other;
-}
-
 class EbookSource{
 	public EbookSourceType $Type;
 	public string $Url;
