+++ lib/EbookSourceType.php
@@ -1,0 +1,12 @@
+<?
+enum EbookSourceType{
+	case ProjectGutenberg;
+	case ProjectGutenbergAustralia;
+	case ProjectGutenbergCanada;
+	case InternetArchive;
+	case HathiTrust;
+	case Wikisource;
+	case GoogleBooks;
+	case FadedPage;
+	case Other;
+}
